Extract user into a context provider
