+++ src/Context/index.ts
@@ -1,2 +1,3 @@
 export { default as AuthContextProvider, useAuthContext } from './AuthContextProvider'
+export { default as UserContextProvider, useUserContext } from './UserContextProvider'
 export { default as WebsocketContextProvider, useWebsocketContext } from './WebsocketContextProvider'
