+++ src/Home/Home.tsx
@@ -1,36 +1,15 @@
 import React from 'react'
-import { useQuery } from '@apollo/react-hooks'
-import { Box, Heading, Flex } from 'rebass'
-import { Cast, Loader, Users } from 'react-feather'
+import { Box, Heading } from 'rebass'
+import { Cast, Users } from 'react-feather'
 
-import { USER_QUERY, UserQuery } from './graphql'
+import Container from 'components/Container'
+import { useUserContext } from 'Context'
+
 import RoomSelector from './RoomSelector'
 import TeamSelector from './TeamSelector'
 
-import Container from '../components/Container'
-
 const Home: React.FC = () => {
-  const { loading, error, data } = useQuery<UserQuery['data']>(USER_QUERY)
-
-  if (loading) {
-    return (
-      <Flex alignItems="center">
-        <Box
-          bg="accent"
-          p={4}
-          sx={{
-            borderRadius: 4,
-          }}
-        >
-          <Loader />
-        </Box>
-      </Flex>
-    )
-  }
-
-  if (!data || error) {
-    return <p>Error</p>
-  }
+  const user = useUserContext()
 
   return (
     <Container>
@@ -41,13 +20,13 @@ const Home: React.FC = () => {
         </Heading>
       </Box>
 
-      <TeamSelector teams={data.user.teams} activeTeam={data.user.activeTeam?.id} />
+      <TeamSelector teams={user.teams} activeTeam={user.activeTeam?.id} />
 
       <Heading as="h2" mb={3} fontSize={[4, 6]}>
         <Cast />
         Available Rooms
       </Heading>
-      {data.user.activeTeam?.id && <RoomSelector activeRoom={data.user.activeRoom?.id} />}
+      {user.activeTeam?.id && <RoomSelector activeRoom={user.activeRoom?.id} />}
     </Container>
   )
 }
